apps/signal-trade/requirements.txt
aiohttp>=3.10,<4
curl-cffi>=0.7,<1
python-dotenv>=1,<2
